Updating some repositories classes

+++ src/main/java/br/com/uninter/sghss/repositories/AgendamentoRepository.java
@@ -3,5 +3,5 @@ package br.com.uninter.sghss.repositories;
 import br.com.uninter.sghss.entities.Agendamento;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AgendamentoRepository extends JpaRepository<Agendamento, Integer> {
+public interface AgendamentoRepository extends JpaRepository<Agendamento, Long> {
 }
