+++ src/main/java/br/com/uninter/sghss/repositories/ConsultaRepository.java
@@ -3,5 +3,5 @@ package br.com.uninter.sghss.repositories;
 import br.com.uninter.sghss.entities.Consulta;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ConsultaRepository extends JpaRepository<Consulta, Integer> {
+public interface ConsultaRepository extends JpaRepository<Consulta, Long> {
 }
