+++ src/main/java/br/com/uninter/sghss/repositories/ExameRepository.java
@@ -3,5 +3,5 @@ package br.com.uninter.sghss.repositories;
 import br.com.uninter.sghss.entities.Exame;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ExameRepository extends JpaRepository<Exame, Integer> {
+public interface ExameRepository extends JpaRepository<Exame, Long> {
 }
