+++ src/main/java/br/com/uninter/sghss/repositories/InternacaoRepository.java
@@ -3,5 +3,5 @@ package br.com.uninter.sghss.repositories;
 import br.com.uninter.sghss.entities.Internacao;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface InternacaoRepository extends JpaRepository<Internacao, Integer> {
+public interface InternacaoRepository extends JpaRepository<Internacao, Long> {
 }
