+++ src/main/java/br/com/uninter/sghss/repositories/LeitoRepository.java
@@ -3,5 +3,5 @@ package br.com.uninter.sghss.repositories;
 import br.com.uninter.sghss.entities.Leito;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface LeitoRepository extends JpaRepository<Leito, Integer> {
+public interface LeitoRepository extends JpaRepository<Leito, Long> {
 }
