+++ src/main/java/br/com/uninter/sghss/repositories/RelatorioFinanceiroRepository.java
@@ -3,5 +3,5 @@ package br.com.uninter.sghss.repositories;
 import br.com.uninter.sghss.entities.RelatorioFinanceiro;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RelatorioFinanceiroRepository extends JpaRepository<RelatorioFinanceiro, Integer> {
+public interface RelatorioFinanceiroRepository extends JpaRepository<RelatorioFinanceiro, Long> {
 }
